Add --limit range validation to location and search commands
- Validate limit is between 1-1000 using clap value_parser
- Change limit type from usize to u32 for range validation support
- Add tests for boundary validation

+++ src/commands/location/execute.rs
@@ -76,7 +76,7 @@ fn find_locations(
     arity: Option<i64>,
     project: Option<&str>,
     use_regex: bool,
-    limit: usize,
+    limit: u32,
 ) -> Result<Vec<FunctionLocation>, Box<dyn Error>> {
     // Build the query based on whether we're using regex or exact match
     let fn_cond = if use_regex {
